+++ config/hypr/scripts/Kool_Quick_Settings.sh
@@ -280,6 +280,7 @@ Configure Workspace Rules (nwg-displays)
 GTK Settings (nwg-look)
 QT Apps Settings (qt6ct)
 QT Apps Settings (qt5ct)
+Choose Hyprlock Wallpaper
 Choose Hyprland Animations
 Choose Monitor Profiles
 Choose Rofi Themes
@@ -366,6 +367,7 @@ main() {
                 exit 1
             fi
             qt5ct ;;
+        "Choose Hyprlock Wallpaper") $scriptsDir/HyprlockWallpaperSelect.sh ;;
         "Choose Hyprland Animations") $scriptsDir/Animations.sh ;;
         "Choose Monitor Profiles") $scriptsDir/MonitorProfiles.sh ;;
         "Choose Rofi Themes") $scriptsDir/RofiThemeSelector.sh ;;